finale/requirements.txt
opencv-python
numpy
pillow
pytesseract
pymupdf
tk
